fix for 64bit.
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@20426 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 3a1c3a7e1030446eb7f2a37d0a4b09d8fe2e93cf

+++ util/geometric_shapes/src/bodies.cpp
@@ -765,7 +765,7 @@ bool bodies::ConvexMesh::intersectsRay(const btVector3& origin, const btVector3&
     if (intersections)
     {
 	std::sort(ipts.begin(), ipts.end(), detail::interscOrder());
-	const unsigned int n = count > 0 ? std::min(count, ipts.size()) : ipts.size();
+	const unsigned int n = count > 0 ? std::min((size_t)count, (size_t)ipts.size()) : ipts.size();
 	for (unsigned int i = 0 ; i < n ; ++i)
 	    intersections->push_back(ipts[i].pt);
     }
